+++ NEWS.md
@@ -1,3 +1,11 @@
+# conformalInference.fd 1.1.0
+
+## version 1.1.0
+
+- Added Jackknife+ extension to functional case
+- modified multisplit with depth measures
+- changed a few inputs for functions
+
 ## version 1.0.0
 
 ---